+++ app/admin/orders/manual/edit/[id]/page.tsx
@@ -160,7 +160,7 @@ export default function EditManualOrderPage({ params }: EditManualOrderPageProps
   const copyPublicLink = () => {
     const url =
       typeof window !== "undefined"
-        ? `${window.location.origin}/order/${order.publicLink}`
+        ? `${window.location.origin}/orders/public/${order.publicLink}`
         : ""
     if (url) {
       navigator.clipboard.writeText(url)
@@ -171,7 +171,7 @@ export default function EditManualOrderPage({ params }: EditManualOrderPageProps
   const openPublicLink = () => {
     const url =
       typeof window !== "undefined"
-        ? `${window.location.origin}/order/${order.publicLink}`
+        ? `${window.location.origin}/orders/public/${order.publicLink}`
         : ""
     if (url && typeof window !== "undefined") {
       window.open(url, "_blank")
