+++ app/admin/orders/manual/page.tsx
@@ -83,7 +83,7 @@ export default function ManualOrdersPage() {
   const copyPublicLink = (publicLink: string) => {
     const url =
       typeof window !== "undefined"
-        ? `${window.location.origin}/order/${publicLink}`
+        ? `${window.location.origin}/orders/public/${publicLink}`
         : ""
     if (url) {
       navigator.clipboard.writeText(url)
@@ -94,7 +94,7 @@ export default function ManualOrdersPage() {
   const openPublicLink = (publicLink: string) => {
     const url =
       typeof window !== "undefined"
-        ? `${window.location.origin}/order/${publicLink}`
+        ? `${window.location.origin}/orders/public/${publicLink}`
         : ""
     if (url && typeof window !== "undefined") {
       window.open(url, "_blank")
